added empty constructors to model classes for serializability by firestore, refactored folder structure to organize
