+++ app/src/main/AndroidManifest.xml
@@ -9,12 +9,12 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ProfileActivity"></activity>
-        <activity android:name=".DashboardActivity" />
-        <activity android:name=".RegisterActivity" />
-        <activity android:name=".MainActivity" />
+        <activity android:name=".ViewModels.ProfileActivity"></activity>
+        <activity android:name=".ViewModels.DashboardActivity" />
+        <activity android:name=".ViewModels.RegisterActivity" />
+        <activity android:name=".ViewModels.MainActivity" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".ViewModels.LoginActivity"
             android:label="@string/title_activity_login"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
