+++ app/src/main/java/com/example/suberduberuber/Models/Request.java
@@ -29,6 +29,9 @@ public class Request {
         changeStatus(status);
     }
 
+    // Empty public constructor needed by Cloud Firestore for serializability
+    public Request() { }
+
     /**
      * Change the status of the request
      * @param newStatus                     The new status of the request, either "initiated", "in-progress", or "completed"
